pci重构测试完成

+++ kernel/src/smp/core.rs
@@ -1,7 +1,7 @@
 /// @brief 获取当前的cpu id
 #[inline]
 pub fn smp_get_processor_id() -> u32 {
-    return crate::arch::cpu::current_cpu_id() as u32;
+    return crate::arch::cpu::current_cpu_id();
 }
 
 #[inline]
